+++ src/components/switch/Switch.vue
@@ -4,18 +4,17 @@
         ref="label"
         :disabled="disabled"
         :tabindex="disabled ? false : 0"
-        @keydown.prevent.enter.space="updateValue(!newValue, $event)"
+        @keydown.prevent.enter.space="$refs.label.click()"
         @mousedown="isMouseDown = true"
         @mouseup="isMouseDown = false"
         @mouseout="isMouseDown = false"
-        @blur="isMouseDown = false"
-        >
-        <input
+        @blur="isMouseDown = false">
+        <input v-model="newValue"
             type="checkbox"
             :name="name"
             :disabled="disabled"
-            v-model="newValue"
-            @change="updateValue(newValue, $event)">
+            :true-value="trueValue"
+            :false-value="falseValue">
         <span class="check" :class="{ 'is-elastic': isMouseDown }"></span>
         <span class="control-label"><slot></slot></span>
     </label>
@@ -25,15 +24,23 @@
     export default {
         name: 'bSwitch',
         props: {
-            value: Boolean,
+            value: {},
+            nativeValue: {},
             disabled: Boolean,
             name: String,
-            checked: Boolean,
-            size: String
+            size: String,
+            trueValue: {
+                type: [String, Number, Boolean, Function, Object, Array, Symbol],
+                default: true
+            },
+            falseValue: {
+                type: [String, Number, Boolean, Function, Object, Array, Symbol],
+                default: false
+            }
         },
         data() {
             return {
-                newValue: this.value || this.checked,
+                newValue: this.value,
                 isMouseDown: false
             }
         },
@@ -50,16 +57,6 @@
              */
             newValue(value) {
                 this.$emit('input', value)
-            }
-        },
-        methods: {
-            /**
-             * Set the newValue.
-             * Emit change event.
-             */
-            updateValue(newValue, $event) {
-                this.newValue = newValue
-                this.$emit('change', newValue, $event)
             }
         }
     }